Migrating logic from parser to compiler.
